Improve template engine with sprig
