+++ core/signature.go
@@ -1,17 +1,14 @@
 package core
 
 import (
-	"bytes"
+	"github.com/jaeles-project/jaeles/database"
+	"github.com/jaeles-project/jaeles/libs"
+	"github.com/jaeles-project/jaeles/utils"
 	"github.com/thoas/go-funk"
 	"path"
 	"path/filepath"
 	"regexp"
 	"strings"
-	"text/template"
-
-	"github.com/jaeles-project/jaeles/database"
-	"github.com/jaeles-project/jaeles/libs"
-	"github.com/jaeles-project/jaeles/utils"
 )
 
 // @NOTE: Signatures allow execute command on your machine
@@ -158,86 +155,4 @@ func AltResolveHeader(headers []map[string]string, target map[string]string) []m
 	}
 
 	return realHeaders
-}
-
-// ResolveVariable resolve template from signature file
-func ResolveVariable(format string, data map[string]string) string {
-	if strings.TrimSpace(format) == "" {
-		return format
-	}
-	_, exist := data["original"]
-	if !exist {
-		data["original"] = ""
-	}
-
-	realFormat, err := template.New("").Parse(format)
-	// when template contain {{
-	if err != nil {
-		r, rerr := regexp.Compile(`\{\{[^.]`)
-		if rerr != nil {
-			return format
-		}
-		matches := r.FindStringSubmatch(format)
-		if len(matches) > 0 {
-			for _, m := range matches {
-				new := strings.Replace(m, `{{`, `{{"{{"}}`, -1)
-				format = strings.Replace(format, m, new, -1)
-			}
-		}
-		// parse it again
-		realFormat, err = template.New("").Parse(format)
-		if err != nil {
-			utils.ErrorF("improper template format %v", format)
-			return format
-		}
-	}
-	t := template.Must(realFormat, err)
-
-	buf := &bytes.Buffer{}
-	err = t.Execute(buf, data)
-	if err != nil {
-		return format
-	}
-	return buf.String()
-}
-
-// AltResolveVariable just like ResolveVariable but looking for [[.var]]
-func AltResolveVariable(format string, data map[string]string) string {
-	if strings.TrimSpace(format) == "" {
-		return format
-	}
-	realFormat, err := template.New("").Delims("[[", "]]").Parse(format)
-	_, exist := data["original"]
-	if !exist {
-		data["original"] = ""
-	}
-
-	// when template contain [[
-	if err != nil {
-		r, rerr := regexp.Compile(`\[\[[^.]`)
-		if rerr != nil {
-			return format
-		}
-		matches := r.FindStringSubmatch(format)
-		if len(matches) > 0 {
-			for _, m := range matches {
-				new := strings.Replace(m, `[[`, `[["[["]]`, -1)
-				format = strings.Replace(format, m, new, -1)
-			}
-		}
-		// parse it again
-		realFormat, err = template.New("").Parse(format)
-		if err != nil {
-			utils.ErrorF("improper template format %v", format)
-			return format
-		}
-	}
-	t := template.Must(realFormat, err)
-
-	buf := &bytes.Buffer{}
-	err = t.Execute(buf, data)
-	if err != nil {
-		return format
-	}
-	return buf.String()
 }
